Improve life quality
As the name of the mod suggests, people install it to have an inventory boost. Setting the default to "0" just makes the process more difficult to use, requiring manual configuration.

The quick-use slots in the middle of the screen don't make for a pleasant experience.

Changes:

Set default number of extra ordinary rows: 2
Set new position of quick slots for better interface experience.